fix windows cli unit imports
